# Remove LinkedDeque and replace with LinkedHashMap