… server/src/main/java/org/opensearch/index/store/remote/utils/cache/Linked.java
… server/src/main/java/org/opensearch/index/store/remote/utils/cache/LinkedDeque.java