@reta 2023-04-04
@andrross I like it, the implementation is much simpler now, interestingly enough, the `LinkedHashMap` supports eviction with `removeEldestEntry` method (that could be implemented) but since we use 2 maps (`data` and `lru` that should be in sync), we may not benefit from it.