@andrross 2023-04-04
@reta regarding `removeEldestEntry`/cache improvements in general, I added some more details here with where the cache has evolved since that issue was opened, with my opinions on what should be improved: https://github.com/opensearch-project/OpenSearch/issues/6225#issuecomment-1496337366