@dblock 2023-05-22
Looks like https://github.com/opensearch-project/OpenSearch/pull/7595 includes this? @kotwanikunal